+++ src/app/feature_modules/user/user-professional-detail-form/user-professional-detail-form.component.html
@@ -57,9 +57,6 @@
               <p-chips formControlName="tags" styleClass="chip-width" [allowDuplicate]='false' [max]=5></p-chips>
             </div>
           </div>
-          <div [ngClass]="['form-row p-2']">
-            <button type="submit" [ngClass]="['btn btn-primary']">Save</button>
-          </div>
         </form>
       </div>
     </div>
