+++ src/app/feature_modules/user/user-work-place-detail-form/user-work-place-detail-form.component.spec.ts
@@ -1,0 +1,25 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { UserWorkPlaceDetailFormComponent } from './user-work-place-detail-form.component';
+
+describe('UserWorkPlaceDetailFormComponent', () => {
+  let component: UserWorkPlaceDetailFormComponent;
+  let fixture: ComponentFixture<UserWorkPlaceDetailFormComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ UserWorkPlaceDetailFormComponent ]
+    })
+    .compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(UserWorkPlaceDetailFormComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
